Search `/usr/include` for CUDA 10.1 headers
As CUDA 10.1 moves some headers and libraries from `/usr/local/cuda` to
`/usr`, we need to adjust our search strategy for finding them. Since
`cudatoolkit` already contains the libraries and is available during the
build, this is less of an issue. However we still need to point the
compiler to the headers in `/usr/include`. Thus we add `/usr/include` to
our compiler flags. To avoid clobbering other search paths we care
about, we make sure to add `/usr/include` last. As we are not searching
`/usr/lib64`, this shouldn't result in any accidental linkages against
system libraries.