Pin to Setuptools pre-72.2.0
In Setuptools 72.2.0, the vendored Distutils was updated. Distutils in
turn pulled in changes for the C++ compiler. This breaks the
configuration build phase of CuPy currently. It is possible to patch and
fix this. However for now just pin Setuptools to an older version that
is known to work. This can be relaxed later when CuPy's build behavior
has been updated.

ref: https://github.com/pypa/setuptools/commit/2c937116cc0dcd9b26b6070e89a3dc5dcbedc2ae