Use only 1 build job in more cases
As seen during `v13` development and testing, the resource usage of
builds has increased. So reducing the number of jobs and/or threads is
needed. Here we choose reducing the number of jobs (as we have done
before) just in more cases. This was found to work during development.
So hopefully it works on the release too.

xref: https://github.com/cupy/cupy/pull/8461