Refresh migrators with global migrator content
This isn't strictly needed as the global migrators will be used unless
`use_local: true` is specified (as was the case before it was dropped in
the previous commit). That said, it is nice to keep these up-to-date.
That way if we later need to vendor them, we can simply add `use_local:
true`. Also it is nice having all the content in one place when
navigating and reading code.